chore refactored DicomTypes to Types so we can include a few types needed for retrieval.
